feat: Moved Navbar and Icon out of components
